프로젝트부분과 마일스톤 부분 개발

+++ src/main/java/com/nhnacademy/taskApi/domain/ProjectStatus.java
@@ -4,6 +4,7 @@ package com.nhnacademy.taskApi.domain;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import javax.persistence.*;
 import java.util.List;
@@ -17,7 +18,7 @@ public class ProjectStatus {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "status_id")
-    private int statusId;
+    private Long statusId;
 
     @Column(name = "status_name")
     private String statusName;
